refactor(test): 클래스를 파일별로 분리
