+++ src/test/kotlin/study/DslTest.kt
@@ -100,74 +100,3 @@ class DslTest {
 fun introduce(block: PersonBuilder.() -> Unit): Person {
     return PersonBuilder().apply(block).build()
 }
-
-class Person(val name: String, val company: String?, val skills: Skills, val languages: Languages)
-
-@JvmInline
-value class Skills(val skills: List<Skill>)
-
-class Skill(val description: String, val type: String)
-
-@JvmInline
-value class Languages(val languages: List<LanguageLevel>)
-
-class LanguageLevel(val country: String, val level: Int)
-
-class PersonBuilder {
-    private lateinit var name: String
-    private var company: String? = null
-    private val skillsList = mutableListOf<Skill>()
-    private val languagesList = mutableListOf<LanguageLevel>()
-
-    fun name(value: String) {
-        name = value
-    }
-
-    fun company(value: String) {
-        company = value
-    }
-
-    fun skills(block: SkillsBuilder.() -> Unit) {
-        val skillsBuilder = SkillsBuilder()
-        skillsBuilder.block()
-        skillsList.addAll(skillsBuilder.skillsList)
-    }
-
-    fun languages(block: LanguagesBuilder.() -> Unit) {
-        val languagesBuilder = LanguagesBuilder()
-        languagesBuilder.block()
-        languagesList.addAll(languagesBuilder.languagesList)
-    }
-
-
-    fun build(): Person {
-        val skills = Skills(skillsList)
-        val languages = Languages(languagesList)
-        return Person(name, company, skills, languages)
-    }
-}
-
-class SkillsBuilder {
-    val skillsList = mutableListOf<Skill>()
-
-    fun soft(description: String) {
-        skillsList.add(Skill(description, SkillType.SOFT.type))
-    }
-
-    fun hard(description: String) {
-        skillsList.add(Skill(description, SkillType.HARD.type))
-    }
-}
-
-enum class SkillType(val type: String) {
-    SOFT("soft"),
-    HARD("hard")
-}
-
-class LanguagesBuilder {
-    val languagesList = mutableListOf<LanguageLevel>()
-
-    infix fun String.level(level: Int) {
-        languagesList.add(LanguageLevel(this, level))
-    }
-}
