+++ src/test/kotlin/study/LanguageLevel.kt
@@ -1,0 +1,3 @@
+package study
+
+class LanguageLevel(val country: String, val level: Int)
