+++ src/test/kotlin/study/Languages.kt
@@ -1,0 +1,4 @@
+package study
+
+@JvmInline
+value class Languages(val languages: List<LanguageLevel>)
