+++ src/test/kotlin/study/LanguagesBuilder.kt
@@ -1,0 +1,9 @@
+package study
+
+class LanguagesBuilder {
+    val languagesList = mutableListOf<LanguageLevel>()
+
+    infix fun String.level(level: Int) {
+        languagesList.add(LanguageLevel(this, level))
+    }
+}
