+++ src/test/kotlin/study/Person.kt
@@ -1,0 +1,3 @@
+package study
+
+class Person(val name: String, val company: String?, val skills: Skills, val languages: Languages)
