+++ src/test/kotlin/study/Skill.kt
@@ -1,0 +1,3 @@
+package study
+
+class Skill(val description: String, val type: String)
