+++ src/test/kotlin/study/SkillType.kt
@@ -1,0 +1,6 @@
+package study
+
+enum class SkillType(val type: String) {
+    SOFT("soft"),
+    HARD("hard")
+}
