+++ src/test/kotlin/study/Skills.kt
@@ -1,0 +1,4 @@
+package study
+
+@JvmInline
+value class Skills(val skills: List<Skill>)
