+++ src/test/kotlin/study/SkillsBuilder.kt
@@ -1,0 +1,13 @@
+package study
+
+class SkillsBuilder {
+    val skillsList = mutableListOf<Skill>()
+
+    fun soft(description: String) {
+        skillsList.add(Skill(description, SkillType.SOFT.type))
+    }
+
+    fun hard(description: String) {
+        skillsList.add(Skill(description, SkillType.HARD.type))
+    }
+}
